+++ assets/jsx/workflow-editor/components/inspector-card/index.jsx
@@ -1,19 +1,125 @@
-import { useSelect } from "@wordpress/data";
+import { useSelect, useDispatch } from "@wordpress/data";
 import { store as editorStore } from "../editor-store";
+import { store as workflowStore } from "../workflow-store";
 import {
     FEATURE_DEVELOPER_MODE,
-    FEATURE_ADVANCED_SETTINGS,
 } from "../../constants";
 import { __ } from "@wordpress/i18n";
-export const InspectorCard = ({ title, description, icon, id, slug, isProFeature }) => {
+import {
+    Button,
+    Popover,
+    TextareaControl,
+    __experimentalVStack as VStack,
+    __experimentalHStack as HStack,
+} from "@wordpress/components";
+import { useState, useEffect } from "@wordpress/element";
+
+const StepDescription = ({ node }) => {
+    const [isPopoverOpen, setIsPopoverOpen] = useState(false);
+    const [description, setDescription] = useState(node.data?.label);
+
+    const {
+        updateNode
+    } = useDispatch(workflowStore);
+
+    useEffect(() => {
+        setIsPopoverOpen(false);
+        setDescription(node.data?.label);
+    }, [node]);
+
+    const closePopover = () => {
+        setIsPopoverOpen(false);
+
+        const newNode = {
+            id: node.id,
+            data: {
+                label: description,
+            },
+        };
+
+        updateNode(newNode);
+    };
+
+
+    return (
+        <>
+            {! node.data?.label && (
+                <Button
+                    variant="link"
+                    onClick={() => {
+                        setIsPopoverOpen(true);
+                    }}
+                >
+                    {__("Add a description to this step...", "post-expirator")}
+                </Button>
+            )}
+
+            {node.data?.label && (
+                <>
+                    <VStack>
+                        <div className="workflow-editor-inspector-card__description">
+                            {node.data.label}
+                        </div>
+
+                        <Button
+                            variant="link"
+                            onClick={() => {
+                                setIsPopoverOpen(true);
+                            }}
+                        >
+                            {__("Edit description", "post-expirator")}
+                        </Button>
+                    </VStack>
+                </>
+            )}
+
+            {isPopoverOpen && (
+                <Popover
+                    onClose={closePopover}
+                    placement="left-start"
+                    offset={80}
+                    className="workflow-editor-inspector-card__description-popover"
+                >
+                    <VStack>
+                        <HStack>
+                            <h2 className="components-truncate components-text components-heading block-editor-inspector-popover-header__heading">
+                                {__("Edit description", "post-expirator")}
+                            </h2>
+                            <Button
+                                icon={'no-alt'}
+                                isSmall={true}
+                                className="block-editor-inspector-popover-header__action"
+                                onClick={closePopover}
+                            />
+                        </HStack>
+                    </VStack>
+
+                    <VStack>
+                        <TextareaControl
+                            value={description}
+                            onChange={(value) => {
+                                setDescription(value);
+                            }}
+                            onKeyDown={(event) => {
+                                if (event.key === 'Escape') {
+                                    closePopover();
+                                }
+                            }}
+                        />
+                    </VStack>
+                </Popover>
+            )}
+        </>
+    );
+}
+
+export const InspectorCard = ({ title, description, icon, id, slug, isProFeature, node }) => {
     const {
         isDeveloperModeEnabled,
-        isAdvancedSettingsEnabled,
         isPro
     } = useSelect((select) => {
         return {
             isDeveloperModeEnabled: select(editorStore).isFeatureActive(FEATURE_DEVELOPER_MODE),
-            isAdvancedSettingsEnabled: select(editorStore).isFeatureActive(FEATURE_ADVANCED_SETTINGS),
             isPro: select(editorStore).isPro(),
         };
     });
@@ -35,32 +141,40 @@ export const InspectorCard = ({ title, description, icon, id, slug, isProFeature
 
     return (
         <div className="workflow-editor-inspector-card">
-            <span className="workflow-editor-inspector-icon has-colors">
-                {icon}
-            </span>
             <div className="workflow-editor-inspector-card__content">
-                <h2 className="workflow-editor-inspector-card__title">
-                    {title}
-                    {isProFeature && !isPro && (
-                        <span className="workflow-editor-inspector-card__pro-badge">
-                            {__("Pro", "post-expirator")}
+                <VStack>
+                    <HStack className="workflow-editor-inspector-card__header">
+                        <span className="workflow-editor-inspector-icon has-colors">
+                            {icon}
                         </span>
-                    )}
-                </h2>
-                <div className="workflow-editor-inspector-card__description">
-                    {description}
-                </div>
+                        <h2 className="workflow-editor-inspector-card__title">
+                            {title}
+                            {isProFeature && !isPro && (
+                                <span className="workflow-editor-inspector-card__pro-badge">
+                                    {__("Pro", "post-expirator")}
+                                </span>
+                            )}
+                        </h2>
+                    </HStack>
+                </VStack>
+                <VStack>
+                    <div className="workflow-editor-inspector-card__description">
+                        {description}
+                    </div>
+                </VStack>
 
                 {isProFeature && !isPro && (
-                    <div className="workflow-editor-inspector-card__pro-instructions">
-                        <a href="https://publishpress.com/links/future-workflow-inspector" target="_blank">
-                        {__("Currently this step is being skipped. Upgrade to Pro to unlock this feature.", "post-expirator")}
-                        </a>
-                    </div>
+                    <VStack>
+                        <div className="workflow-editor-inspector-card__pro-instructions">
+                            <a href="https://publishpress.com/links/future-workflow-inspector" target="_blank">
+                            {__("Currently this step is being skipped. Upgrade to Pro to unlock this feature.", "post-expirator")}
+                            </a>
+                        </div>
+                    </VStack>
                 )}
 
                 {nodeAttributes.length > 0 && (
-                    <>
+                    <VStack>
                         <table>
                             <tbody>
                                 {nodeAttributes.map((attribute) => {
@@ -73,8 +187,12 @@ export const InspectorCard = ({ title, description, icon, id, slug, isProFeature
                                 })}
                             </tbody>
                         </table>
-                    </>
+                    </VStack>
                 )}
+
+                <VStack>
+                    <StepDescription node={node} />
+                </VStack>
             </div>
         </div>
     );